Added the ability to extract feature names from Pipeline
There is no way to get the list of features produced
by the transforming pipeline, so it is hard to analyze
numpy ndarray passed to the model.

Wrapped all the transforms with custom classes that
derive feature names produced by the transformations.
eli5.transform_feature_names() was considered but it didn't match
because it doesn't work out-of-the-box and doesn't support in_names
parameter, producing automatically generated names.

Added feature importance analysis.

Added is_rush_hour feature.